+++ ccc-dns.js
@@ -143,7 +143,7 @@ module.exports = {
 
         // Set default response
         let cdnResponse = {
-            message: 'Unknown',
+            message: [],
             hostname: hostname,
             reason: '',
             service: 'Unknown',
@@ -158,8 +158,6 @@ module.exports = {
                 type: CCC_DNS_REQUEST_RECORD_TYPE,
             });
 
-            //let resolverAddress = module.exports.getDNSResolver();
-            //console.log('DNS Resolver: %s', resolverAddress);
             let req = dns.Request({
                 question: question,
                 server: {address: module.exports.getDNSResolver(), port: 53, type: 'udp'},
@@ -168,7 +166,7 @@ module.exports = {
 
             req.on('timeout', () => { // DNS timeout
                 debug('DNS timeout occurred resolving [%s]', hostname);
-                cdnResponse.message = 'DNS Timeout';
+                cdnResponse.message.push('DNS Timeout');
                 cdnResponse.reason = 'DNS Timeout';
                 cdnResponse.error = `Timeout after ${req.timeout} milliseconds`;
                 cdnResponse.status = CCC_CDN_DETERMINATION_STATUS.ERROR;
@@ -178,7 +176,7 @@ module.exports = {
             req.on('message', (error, answer) => { // DNS message received
                 if (error) { // DNS returned an error
                     debug('Received DNS error for %s: %O', hostname, error);
-                    cdnResponse.message = 'DNS Error';
+                    cdnResponse.message.push('DNS Error');
                     cdnResponse.reason = 'DNS Error';
                     cdnResponse.error = error;
                     cdnResponse.status = CCC_CDN_DETERMINATION_STATUS.ERROR;
@@ -200,7 +198,7 @@ module.exports = {
                                 cdnResponse.reason = `${hostname} resolves to ${matchingDomains[0]} which matches a ${cdn} domain pattern`;
                                 cdnResponse.matchingDomains = matchingDomains[0];
                                 cdnResponse.service = apexDomains[cdn].service.toUpperCase();
-                                cdnResponse.message = apexDomains[cdn].title;
+                                cdnResponse.message.push(apexDomains[cdn].title);
                                 cdnResponse.status = CCC_CDN_DETERMINATION_STATUS.CDN;
                                 if (apexDomains[cdn].service.toUpperCase() !== 'CDN') {
                                     cdnResponse.status = CCC_CDN_DETERMINATION_STATUS.OTHER;
@@ -220,7 +218,7 @@ module.exports = {
                         let awsServicesFile = __dirname + pathSeparator + 'service.providers/aws/ip-ranges.json';
                         let rawData = fs.readFileSync(awsServicesFile); // Read the AWS services file
                         let awsServices = JSON.parse(rawData); // Parse it into a JSON object
-                        let awsServicesMessage = ''; // Temporarily store the message because the AWS JSON might contain two matching CIDR blocks, so we can't just concatenate
+                        let awsServicesMessage = []; // Temporarily store the message because the AWS JSON might contain two matching CIDR blocks, so we can't just concatenate
 
                         // Loop through each service
                         debug('Checking if the IP address [%s] matches a known AWS service', cdnResponse.ipAddress);
@@ -231,41 +229,31 @@ module.exports = {
                             // Check if the IP address exists within the cidr block
                             if (cidr.contains(cdnResponse.ipAddress)) {
                                 debug('%s is in the CIDR block %s, which is AWS service %s', cdnResponse.ipAddress, awsServices.prefixes[i].ip_prefix, awsServices.prefixes[i].service);
-                                awsServicesMessage = awsServices.prefixes[i].service;
+                                awsServicesMessage.push(awsServices.prefixes[i].service);
                                 cdnResponse.status = CCC_CDN_DETERMINATION_STATUS.AWS;
                                 cdnResponse.reason = `${cdnResponse.ipAddress} is in the CIDR block ${awsServices.prefixes[i].ip_prefix} which is used by AWS ${awsServices.prefixes[i].service}`;
 
-                                if (String.prototype.toUpperCase.call(awsServicesMessage) === 'CLOUDFRONT') { // Check if the service is CloudFront
-                                    awsServicesMessage = 'CloudFront';
+                                if (String.prototype.toUpperCase.call(awsServicesMessage[awsServicesMessage.length - 1]) === 'CLOUDFRONT') { // Check if the service is CloudFront
                                     cdnResponse.service = 'CDN';
                                     cdnResponse.status = CCC_CDN_DETERMINATION_STATUS.CDN;
-                                } else {
-                                    cdnResponse.status = CCC_CDN_DETERMINATION_STATUS.AWS;
-                                    cdnResponse.service = awsServices.prefixes[i].service;
-                                    if (String.prototype.toUpperCase.call(awsServices.prefixes[i].region) != 'GLOBAL') { // Append the region if it's not ambiguous
-                                        awsServicesMessage += ' (' + awsServices.prefixes[i].region + ')';
-
-                                    }
                                 }
                             }
                         }
 
-                        if (cdnResponse.status === CCC_CDN_DETERMINATION_STATUS.AWS) {
+                        if (awsServicesMessage.length > 0) {
                             // Save the generated message into the response object
-                            cdnResponse.message = awsServicesMessage;
-
-                            /* if (cdnResponse.message === 'Unknown') {
-                                debug('Replacing the message [%s] with [%s]', cdnResponse.message, awsServicesMessage);
-                                cdnResponse.message = awsServicesMessage; // Replace the current message with a more precise AWS specific detection
-                            } else {
-                                debug('Replacing the message [%s] with [%s]', cdnResponse.message, cdnResponse.message += ', ' + awsServicesMessage);
-                                cdnResponse.message += ', ' + awsServicesMessage; // Append the AWS detection message to the DNS detection message
-                            } */
+                            cdnResponse.message.push('[' + awsServicesMessage.join(' -> ') + ']');
                         }
 
                     }
 
                     debug('determineCDN(%s) returning: %O', hostname, cdnResponse);
+
+                    // Check if the message array is blank, and add the default message if it is
+                    if (cdnResponse.message.length === 0) {
+                        // We didn't identify the service behind the domain name or IP address
+                        cdnResponse.message = ['Unknown'];
+                    }
                     callback(cdnResponse);
                 }
             });